+++ distributions/nuxt/package.json
@@ -53,8 +53,8 @@
     "@nuxt/common": "2.3.4",
     "@nuxt/core": "2.3.4",
     "@nuxt/generator": "2.3.4",
-    "@nuxt/webpack": "2.3.4",
-    "@nuxt/opencollective": "^0.2.1"
+    "@nuxt/opencollective": "^0.2.1",
+    "@nuxt/webpack": "2.3.4"
   },
   "engines": {
     "node": ">=8.0.0",
