+++ packages/builder/package.json
@@ -8,9 +8,9 @@
   ],
   "main": "dist/builder.js",
   "dependencies": {
-    "@nuxt/common": "2.3.4",
-    "@nuxt/vue-app": "2.3.4",
     "@nuxt/devalue": "^1.2.0",
+    "@nuxt/utils": "2.3.4",
+    "@nuxt/vue-app": "2.3.4",
     "chokidar": "^2.0.4",
     "consola": "^2.3.0",
     "fs-extra": "^7.0.1",
