+++ packages/common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@nuxt/common",
+  "name": "@nuxt/utils",
   "version": "2.3.4",
   "repository": "nuxt/nuxt.js",
   "license": "MIT",
@@ -8,8 +8,8 @@
   ],
   "main": "dist/common.js",
   "dependencies": {
-    "consola": "^2.3.0",
-    "serialize-javascript": "^1.5.0"
+    "@nuxt/utils": "2.3.4",
+    "consola": "^2.3.0"
   },
   "publishConfig": {
     "access": "public"
