+++ packages/common/src/index.js
@@ -1,3 +1,2 @@
 export { default as Hookable } from './hookable'
 export { default as ModernBrowsers } from '../data/modern-browsers.json'
-export * from './utils'
