+++ packages/core/package.json
@@ -10,9 +10,10 @@
   "dependencies": {
     "@nuxt/common": "2.3.4",
     "@nuxt/config": "2.3.4",
-    "@nuxt/server": "2.3.4",
-    "@nuxt/vue-renderer": "2.3.4",
     "@nuxt/devalue": "^1.2.0",
+    "@nuxt/server": "2.3.4",
+    "@nuxt/utils": "2.3.4",
+    "@nuxt/vue-renderer": "2.3.4",
     "consola": "^2.3.0",
     "debug": "^4.1.0",
     "esm": "^3.0.84",
