+++ packages/core/src/nuxt.js
@@ -2,7 +2,8 @@
 import isPlainObject from 'lodash/isPlainObject'
 import consola from 'consola'
 
-import { Hookable, defineAlias } from '@nuxt/common'
+import { Hookable } from '@nuxt/common'
+import { defineAlias } from '@nuxt/utils'
 import { getNuxtConfig } from '@nuxt/config'
 import { Server } from '@nuxt/server'
