+++ packages/server/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "@nuxt/common": "2.3.4",
     "@nuxt/config": "2.3.4",
+    "@nuxt/utils": "2.3.4",
     "@nuxtjs/youch": "^4.2.3",
     "chalk": "^2.4.1",
     "compression": "^1.7.3",
