+++ packages/server/src/server.js
@@ -4,7 +4,7 @@ import launchMiddleware from 'launch-editor-middleware'
 import serveStatic from 'serve-static'
 import servePlaceholder from 'serve-placeholder'
 import connect from 'connect'
-import { determineGlobals, isUrl } from '@nuxt/common'
+import { determineGlobals, isUrl } from '@nuxt/utils'
 
 import ServerContext from './context'
 import renderAndGetWindow from './jsdom'
