+++ packages/webpack/package.json
@@ -11,8 +11,8 @@
     "@babel/core": "^7.2.2",
     "@babel/polyfill": "^7.2.3",
     "@nuxt/babel-preset-app": "2.3.4",
-    "@nuxt/common": "2.3.4",
     "@nuxt/friendly-errors-webpack-plugin": "^2.4.0",
+    "@nuxt/utils": "2.3.4",
     "babel-loader": "^8.0.4",
     "cache-loader": "^1.2.5",
     "caniuse-lite": "^1.0.30000923",
