+++ packages/webpack/src/utils/postcss.js
@@ -5,7 +5,7 @@ import merge from 'lodash/merge'
 import cloneDeep from 'lodash/cloneDeep'
 import createResolver from 'postcss-import-resolver'
 
-import { isPureObject } from '@nuxt/common'
+import { isPureObject } from '@nuxt/utils'
 
 export const orderPresets = {
   cssnanoLast: (names) => {
